fix breakage in the perl debugger on perl >= v5.22.0 (#9)
bareword invocants aren't autoboxed e.g.:

    Foo->VERSION # not autoboxed
    "foo"->uc    # OK

previously the bareword test checked to see if a flag was set on
the invocant OP, namely OPpCONST_BARE (0x40). however, this flag
only indicates that the OP is a bareword if the OP is OP_CONST
(it doesn't make sense for other OPs). on an OP_ENTERSUB, it means
OPpENTERSUB_DB, which is set on ENTERSUB OPs inside the debugger.
because this flag was being misinterpreted as "bareword invocant"
rather than "sub call in the debugger" in the OP-check function,
these OPs were being incorrectly exempted from autoboxing.

the bug was exposed by this commit (included in v5.22.0):

    https://github.com/Perl/perl5/commit/9e20940244

which makes better use of the limited number of bits
available for OP flags by reshuffling/reorganizing flags.
this resulted in OPpENTERSUB_DB changing from 0x10 to 0x40,
which exposed this bug.

---

also fix bareword detection for OP_METHOD i.e. Foo->$bar (#8)