allow for other debugger output when checking for expected string
debugger.t expects a particular string to be output when test code is
run under the debugger. It assumes that that will be the only output.

However, the user's .perldb file may generate output,
contaminating the output and thus invalidating the test.

Instead, look for the presence of the expected string in the output,
ignoring everything else.